Begin grand restructuring of events (This commit is non-functional)
