+++ src/main/java/com/tcvcog/tcvce/application/BackingBeanUtils.java
@@ -75,7 +75,6 @@ import java.sql.SQLException;
 
 import java.time.LocalDateTime;
 import java.time.ZoneId;
-import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.Iterator;
